Refactor admin preview into site/block/common subfolders for better logical separation of the two different but similar features
