+++ demo/admin/src/pages/EditPage.tsx
@@ -15,7 +15,7 @@ import {
     BlockPreviewWithTabs,
     createUsePage,
     EditPageLayout,
-    openPreviewWindow,
+    openSitePreviewWindow,
     PageName,
     useBlockPreview,
     useCmsBlockContext,
@@ -157,7 +157,7 @@ export const EditPage: React.FC<Props> = ({ id, category }) => {
                         startIcon={<Preview />}
                         disabled={!pageState}
                         onClick={() => {
-                            openPreviewWindow(pageState.path, contentScopeMatch.url);
+                            openSitePreviewWindow(pageState.path, contentScopeMatch.url);
                         }}
                     >
                         <FormattedMessage id="cometDemo.pages.pages.page.edit.preview" defaultMessage="Web preview" />
