+++ demo/admin/src/pages/mainMenu/components/EditMainMenuItem.tsx
@@ -14,7 +14,7 @@ import {
 } from "@comet/admin";
 import { Add, Delete, Preview, Save } from "@comet/admin-icons";
 import { AdminComponentRoot, BlockOutputApi, BlockState, HiddenInSubroute, IFrameBridgeProvider, resolveNewState } from "@comet/blocks-admin";
-import { EditPageLayout, openPreviewWindow, SplitPreview, useBlockPreview, useCmsBlockContext, useSiteConfig } from "@comet/cms-admin";
+import { EditPageLayout, openSitePreviewWindow, SplitPreview, useBlockPreview, useCmsBlockContext, useSiteConfig } from "@comet/cms-admin";
 import { Button } from "@mui/material";
 import { RichTextBlock } from "@src/common/blocks/RichTextBlock";
 import { useContentScope } from "@src/common/ContentScopeProvider";
@@ -133,7 +133,7 @@ const EditMainMenuItem: React.FunctionComponent<EditMainMenuItemProps> = ({ item
                         color="info"
                         startIcon={<Preview />}
                         onClick={() => {
-                            openPreviewWindow(item.node.path, contentScopeMatch.url);
+                            openSitePreviewWindow(item.node.path, contentScopeMatch.url);
                         }}
                     >
                         <FormattedMessage id="cometDemo.pages.pages.page.edit.preview" defaultMessage="Web preview" />
