+++ packages/admin/cms-admin/src/index.ts
@@ -63,13 +63,14 @@ export { useCopyPastePages } from "./pages/pageTree/useCopyPastePages";
 export { resolveHasSaveConflict } from "./pages/resolveHasSaveConflict";
 export { useSaveConflict } from "./pages/useSaveConflict";
 export { useSaveConflictQuery } from "./pages/useSaveConflictQuery";
-export { BlockPreview } from "./preview/BlockPreview";
-export { BlockPreviewWithTabs } from "./preview/BlockPreviewWithTabs";
-export { openPreviewWindow } from "./preview/openPreviewWindow";
-export { SitePreview } from "./preview/SitePreview";
-export { SplitPreview } from "./preview/SplitPreview";
-export type { BlockPreviewApi } from "./preview/useBlockPreview";
-export { useBlockPreview } from "./preview/useBlockPreview";
+export { BlockPreview } from "./preview/block/BlockPreview";
+export { BlockPreviewWithTabs } from "./preview/block/BlockPreviewWithTabs";
+export { SplitPreview } from "./preview/block/SplitPreview";
+export type { BlockPreviewApi } from "./preview/block/useBlockPreview";
+export { useBlockPreview } from "./preview/block/useBlockPreview";
+export { openSitePreviewWindow } from "./preview/openSitePreviewWindow";
+export { openSitePreviewWindow as openPreviewWindow } from "./preview/openSitePreviewWindow"; //deprecated
+export { SitePreview } from "./preview/site/SitePreview";
 export { createRedirectsPage } from "./redirects/createRedirectsPage";
 export { automaticRedirectsRefetchQueryDescription as automaticRedirectsRefetchQueryDescription__temporary__export } from "./redirects/RedirectsTable.gql";
 export type { SiteConfig } from "./sitesConfig/SitesConfigContext";
