+++ packages/admin/cms-admin/src/preview/common/icons/DeviceFrameDesktop.tsx
