+++ packages/admin/cms-admin/src/preview/common/icons/DeviceFrameMobile.tsx
