+++ packages/admin/cms-admin/src/preview/common/icons/DeviceFrameTablet.tsx
